+++ tests/test_EEE/test_tractor.py
@@ -34,6 +34,7 @@ def test_herd_size_to_tractor_size(
     tractor_dataset: dict[str, list[Any]],
     mocker: MockFixture,
 ) -> None:
+    """Tests the herd size to tractor size conversion."""
     im = InputManager()
     mocker.patch.object(im, "get_data", side_effect=[deepcopy(EEE_constants), tractor_dataset, deepcopy(EEE_constants)])
     specs = Tractor(
@@ -52,17 +53,20 @@ def test_tractor_size_initialization(
     tractor_dataset: dict[str, list[Any]],
     mocker: MockFixture,
 ) -> None:
+    """Tests the tractor size initialization."""
     specs = mock_tractor(tractor_size, EEE_constants, tractor_dataset, mocker)
     assert specs.tractor_size == tractor_size
 
 
 def test_raises_value_error_for_missing_parameters() -> None:
+    """Tests that a ValueError is raised when missing required parameters."""
     with pytest.raises(ValueError):
         Tractor(operation_event=FieldOperationEvent.TILLING)
 
 
 @pytest.mark.parametrize("herd_size", [-1, -100])
 def test_herd_size_negative_value_error(herd_size: int) -> None:
+    """Tests that a ValueError is raised when herd size is negative."""
     with pytest.raises(ValueError):
         Tractor(operation_event=FieldOperationEvent.TILLING, herd_size=herd_size)
 
@@ -139,6 +143,7 @@ def test_determine_operation_type(
     tractor_dataset: dict[str, list[Any]],
     mocker: MockFixture,
 ) -> None:
+    """Tests the operation type determination."""
     im = InputManager()
     mocker.patch.object(im, "get_data", side_effect=[deepcopy(EEE_constants)])
     mocker.patch("RUFAS.EEE.tractor_implement.TractorImplement.__init__", return_value=None)
@@ -168,6 +173,7 @@ def test_pto_kW(
     tractor_dataset: dict[str, list[Any]],
     mocker: MockFixture,
 ) -> None:
+    """Tests the PTO kW calculation."""
     specs = mock_tractor(tractor_size, EEE_constants, tractor_dataset, mocker)
     assert specs.PTO_kW == expected_pto_kW
 
@@ -187,6 +193,7 @@ def test_power_available_kW(
     tractor_dataset: dict[str, list[Any]],
     mocker: MockFixture,
 ) -> None:
+    """Tests the power available kW calculation."""
     specs = mock_tractor(tractor_size, EEE_constants, tractor_dataset, mocker)
     assert specs.power_available_kW == expected_power_available_kW
 
@@ -206,6 +213,7 @@ def test_mass_kg(
     tractor_dataset: dict[str, list[Any]],
     mocker: MockFixture,
 ) -> None:
+    """Tests the mass kg calculation."""
     specs = mock_tractor(tractor_size, EEE_constants, tractor_dataset, mocker)
     assert specs.mass_kg == expected_mass_kg
 
@@ -213,6 +221,7 @@ def test_mass_kg(
 def test_speed_km_hr(
     EEE_constants: list[dict[str, Any]], tractor_dataset: dict[str, list[Any]], mocker: MockFixture
 ) -> None:
+    """Tests the speed km/hr calculation."""
     specs = mock_tractor(TractorSize.SMALL, EEE_constants, tractor_dataset, mocker)  # Any tractor size would do
     assert specs.speed_km_hr == 10.0
 
@@ -241,6 +250,7 @@ def test_calculate_axel_power(
     tractor_dataset: dict[str, list[Any]],
     mocker: MockFixture,
 ) -> None:
+    """Tests the axel power calculation."""
     tractor = mock_tractor(TractorSize.SMALL, EEE_constants, tractor_dataset, mocker)
 
     mocker.patch.object(Tractor, "mass_kg", new_callable=mocker.PropertyMock, return_value=tractor_mass)